Set the TERM variable conditionally in vimrc
When the TERM variable is set to xterm-256colors, there are
transparency issues when using vim in a TMUX session. If set to
screen-256colors, the HOME and END keys are remapped and cannot be used
when using vim outside of a TMUX session.

This change sets the TERM variable based on whether vim is being used
inside a TMUX session.